GASNET_MAX_SEGSIZE changes
Added some logic to chpl_cfg_comm.c so that if we're running on
an XT and GASNET_MAX_SEGSIZE is not set, we set it to what we
believe will be a reasonable value based on /proc/meminfo.
Unfortunately, setting it to the full amount of memory in
/proc/meminfo results in the program terminating with:

        _pmii_daemon(SIGCHLD): PE 0 exit signal Killed

So after some experimentation, I am now subtracting 0.5GB from
the amount returned.  My hypothesis is that this may be required
to cover for our code size, static data, and/or allocations done
outside of the GASNet segment, but I'm not sure yet and hope to
look into it more after this commit.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15396 3a8e244f-b0f2-452b-bcba-4c88e055c3ca