Add a build step to make GASNet installs for XT movable
After the make install step when building GASNet targeting XT, add a step
to replace absolute paths in the GASNet makefile fragments with a variable
we define in our Makefiles before including the fragments.

I only change it when the target is XT, and only changed the portals-par and
mpi-par fragments. This should make it possible to move the entire chapel
directory after GASNet has been 'make install'ed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15405 3a8e244f-b0f2-452b-bcba-4c88e055c3ca