Expanded and edited remaining of David's tutorial examples.  I think I made minor improvements to all of these examples and expanded most of them slightly.  There is certainly more that we could do in the future, but this is a great start.
I changed the timers example to get rid of the need for a magic number
since some systems report that a call to sleep(1) takes less than a
second.  We already have a test that makes sure timers report
something reasonable.  See trivial/deitz/test_timer1.chpl, so I don't
think the old version of this test, with the magic number, are unique.
I think actually reporting the time, and using a config const for
testing, makes the test simpler for the tutorial purposes.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15426 3a8e244f-b0f2-452b-bcba-4c88e055c3ca