Updates to IBM launcher
Based on last-minute testing, determined that the loadleveler launcher
is not really stable enough to serve as the default for pwr5/pwr6
platforms for this release.  Disabled it as the default in
launcher.pl, updated README.launcher to reflect the change, and added
more information to README.ibm to give people a handle on how to get
things running (along with a request for help from IBM enthusiasts to
help us get a better loadleveler launcher written).

Also added a note to README.ibm about the fact that our memory query
does not seem to be portable to the power5.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15458 3a8e244f-b0f2-452b-bcba-4c88e055c3ca