Fixed several baseline regressions involving double freeing or freeing strings that were not malloc'ed.  The problem was that if a tuple contained a string, the destructor for the tuple would free the string. However, when strings are put into tuples, they are not actually copied, so destructors for tuples should not free strings.  This is not normally a problem because inlining and scalar replacement changes the code for tuples, which eliminates the need to call destructors for tuples, at least those containing strings.
The fix is to not free strings in destructors for tuples.  This will result in
a few new baseline regressions because strings may be leaked in cases in which
they were not previously being leaked.  However, these leaks should not be
eliminated by (mistakenly) freeing strings in destructors for tuples.  Rather,
better analysis needs to be done so that when a string that was malloc'ed
(due to calling string_copy) is put into a tuple, the lifetime of the tuple
is tracked along with the lifetime of the malloc'ed string, which typically
ends when it is copied into the tuple.  Then, when the tuple is no longer being
used (or goes out of scope), the string can be freed separately (not in the
destructor for the tuple).

An alternative would be to actually copy the string when putting it into the
tuple.  However, note that this is probably not a good idea when dealing with
arrays, for example.  So the above analysis would still need to be implemented,
at least for the case of arrays.  Anyway, this analysis is not very difficult
to implement because handling a very specific idiom in which an object is
allocated in the same block in which the tuple to which it is assigned is
declared and used would likely take care of 99% of leaks of this kind.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15462 3a8e244f-b0f2-452b-bcba-4c88e055c3ca