This change updates the PVM comm layer with recent changes.
This change updates the PVM comm layer with recent changes, though it does not represent one complete block of functionality. The launcher now explicity spawns for each numLocale (list from hostfile). The comm layer framework is very similar to that which is found in the MPI layer, though a lot of work needs to be done. Currently, the pthread worker is the main focus. Since there is nothing analogous to MPI_Status in PVM, it would make sense to package the message_info information with the send/receive. The next thing we'll need here is a type enumeration. Note that until we prevent the PVM shell from quitting as soon as the commands are launched, the user has to manually start the _real app interactively (with the right flags).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15465 3a8e244f-b0f2-452b-bcba-4c88e055c3ca