2nd attempt at First sourceforge commit: re-adding quickReference
If this works, it is our first commit after the conversion to
SourceForge and will also be an initial test of the hook script mail.

In making the conversion to Sourceforge, we eliminated the
doc/internal directory which contained (among other things that we
didn't want to move into public) the quickReference files. Here I am
re-adding the current version of those files.

The other things that were filtered out in the conversion were a few
potentially sensitive user codes: test/users/wmikanik/ca* and
test/users/scott. Better safe than sorry.

Fingers crossed.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15472 3a8e244f-b0f2-452b-bcba-4c88e055c3ca