In the int(64) op uint(64) operators don't use primitives directly.
Instead, cast the int(64) to uint(64) and use the uint(64) versions of the
operators. This way, if the implementation of the operator with same types for
arguments changes, these automatically follow the change. It is still a
compile time error to call these int(64) op uint(64) operators with a
non-parameter or negative int(64).

Updated the Chapel util "script" used to generate these.

This should clean up some Intel compiler warnings about calling the C pow
function on uint(64) values.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15479 3a8e244f-b0f2-452b-bcba-4c88e055c3ca