Add a future about calling a bool method from a varags function.
This test declares a method over the bool type, then attempts to call it on
one of the arguments in a variable length argument list. It currently results
in ambiguous call error with the method listed for each bool size. If the
argument is copied into a temp before calling the method, it resolves to a
single instance.

The bool method I would like to write and call in this way is
bool.writeThis(f: Writer) in the ChapelIO module.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15492 3a8e244f-b0f2-452b-bcba-4c88e055c3ca