Added an extra argument to printMemTable, printToMemLog, installMemory, increaseMemStat, and decreaseMemStat to keep track of whether memory allocations are due to explicit user code or not.  Also, made the memory usage reporting default to tracking only allocations explicit in user code, unless --memthreshold is set to a negative number, in which case only allocations unrelated to explicit user code is reported.  (The total amount of user and nonuser memory allocated is still tracked regardless in order to accurately report when memory usage is above the given limit.)  The given threshold is now the smallest size for allocations that will be reported, so, for example, a threshold of -4 reports only nonuser allocations of 4 bytes or more.  (Perhaps it would be good to report both user and nonuser allocations if the threshold is 0, but this is not implemented in this check in.)
Also made the output file used for --memtrace be stdout if the file name is "-",
and made chpl_startMemDiagnosis and chpl_stopMemDiagnosis applicable only if
the macro MEM_DIAGNOSIS in chplmem.c is set to a nonzero value and --memtrace
is specified at run time.

Finally, made several code safety improvements to ensure correct behavior in
certain edge cases that should never occur, and made a number of formatting
changes.

Two test cases no longer succeed; they were turned into futures.  One of them,
test/memory/shannon/printMemTable2.chpl, doesn't work because it allocates new
instances of a class, but there is not yet a way to propagate the fact that
these allocations occurred because of explicit user code.  The other one,
test/memory/shannon/allocAndFree.chpl, doesn't work because there is not yet a
way to propagate the fact that a call to chpl_free (or calls to destructors)
are due to explicit user code.  In addition, it contains a call to chpl_realloc,
and the only other place that uses this function is in chplcomm.c, so at least
currently, there is no need for chpl_realloc to keep track of whether it was
called due to explicit user code, which would be required for this specific
test case to work properly.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15495 3a8e244f-b0f2-452b-bcba-4c88e055c3ca