Remove automatic coercions from numeric, bool, and enum types to string.
Overloaded the '+' operator for strings and numeric types, for param strings
and param integral types, for strings and enums and for param strings and
param enums.

Added writeThis methods for numeric, bool and enumerated types which cast
"this" to string. The bool.writeThis method is currently commented out due
to the problem shown in test/functions/diten/boolMethodVarArg.chpl. As a
workaround, the Writer.write method checks for bools and uses the strings
"true" and "false" instead of calling bool.writeThis.

Fold casts from param enumeration values to string.

Remove casts that don't change the type of the the value they act on. This
was required for test/modules/standard/Time/stonea/getCurrentDayOfWeek.chpl.
This test casts an enum value of type Day to type Day. This cast doesn't
exist and instead the value was implicitly cast to string and the explicit
cast to type Day then converted the string back into an enum Day.

Update a few .good files where changing the output of the test seems more
appropriate than explicitly casting values to string and one .good where
line numbers changed in an error message.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15499 3a8e244f-b0f2-452b-bcba-4c88e055c3ca