Initial attempt to propagate whether an allocation is due to explicit user code.
A previous check in added an extra argument to calls to the primitive PRIM_NEW.
This check in passes that extra argument to the resulting call to the
corresponding constructor if this argument is true.  (This gets around a problem
during function resolution for constructors that don't have this extra
argument.)

In order for this to work, an extra formal argument called "userCode" is
inserted in constructors just before the meme formal.  The default value for
userCode is false, which indicates the call is not due to user explicit code.
When a call to chpl_alloc is added to wrappers for constructors, the userCode
argument is passed along, if it is one of the arguments to the wrapper.

There are two problems with this check in, which, if serious enough, could
warrant backing out the changes.  One is that the extra userCode argument is not
propagated to any constructors that might get called from within a constructor
that has this extra argument.

The other problem is that error messages resulting from unresolved calls to
constructors will have this extra argument, even though there is no Chapel code
that refers to this argument.  This could be confusing to Chapel users.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15512 3a8e244f-b0f2-452b-bcba-4c88e055c3ca