minor XMT improvements
I thought I had cleaned out all the XMT-specific stdint.h files
but Sam's checkin the other week showed that this was not the
case in the runtime.  Fortunately, these ones are not necessary
any longer, so I'm removing them here.

I also added a warning suppression for the chapel_code.c file
in the generated code to avoid warnings relating to possible
full/empty bit modifications on non-word-sized data.  As with
the normal generated code, the theory is that we've padded
things approptiately and guaranteed the safety of this, so the
warnings can be safely ignored.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15517 3a8e244f-b0f2-452b-bcba-4c88e055c3ca