Adding pvm 3.4.6 to our third-party directory
I'm adding pvm 3.4.6 to our third-party directory for use
in the hybrid computing project.  This is essentially just
a snapshot of the 3.4.6 release with some minor Makefile
and README changes to document its addition.  Next step is
to make the launch-pvm and comm-pvm directories build
against this version of pvm at which point it should star
to become useful.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15521 3a8e244f-b0f2-452b-bcba-4c88e055c3ca