Added a check to our handling of module access expressions to make sure that the symbols in the accessed module are visible via a use or directly.  Removed future for test/modules/bradc/configParamNoUse and trivial/deitz/test_module_access1.  Removed 'use Random' from a routine in the Random module which was necessary since this routine could be instantiated outside of the Random module.  Now, the Random module needs to be used whereever this routine can be called.  Updated the test of fillRandom to report an error.  Updated several other tests by changing the .good to report an error or by inserting a call to use a module before the module access expression.
Fixed a bug in the _copy function defined for records in which it used
to call the default constructor by name; it now calls the default
constructor explicitly.  This is necessary because user-defined
constructors shadow the default one.  Removed future for
classes/stonea/recordConstructor.

Fixed a bug in the parameter cast code in which the guarding condition
used to allow anything other than real/imag/complex types through even
if it would not handle things other than int/uint/bool/string.  I
modified the condition so that other things skip the folding.  The
error message for the test is then appropriate and is the same as
non-parameter casts.  Removed future for param/diten/castParamToClass.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15529 3a8e244f-b0f2-452b-bcba-4c88e055c3ca