Fixing GASNet regressions due to -nl x throwing off argument counting
In last night's checkin, I failed to anticipate that our prepending of
the -nl x flag on GASNet runs would throw off the argument counting
used by our error messages, resulting in a few dozen regressions.

Here, I'm fixing this by moving the -nl x argument for the test system
to the end of the command line instead of putting it at the beginning.
For some tests (the ones listed explicitly and in the directories
below) this still resulted in an error because the tests were
explicitly checking things like "what if the command line ends before
we find the value of a flag?"  However, this number of tests was much
smaller than the number of regressions last night, and the fix for
them was easy: adding NUMLOCALES or .numlocales files with "0" as the
number of locales causes the test system not to add a -nl option at
all, so I did this for these tests that are explicitly testing command
line parsing errors.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15534 3a8e244f-b0f2-452b-bcba-4c88e055c3ca