Improved parsing of numLocales flag
While working on adding realms support, realized that numLocales was
not being treated as similarly to other config vars/consts as it
should be.  This check-in makes it less of a special case in our
argument passing code and adds a test ensuring that it can be set
using --numLocales 2 as with other config vars.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15542 3a8e244f-b0f2-452b-bcba-4c88e055c3ca