Changed how memory leaks are reported when --memfinalstat is used at run time. When a program finishes running, if there are no memory leaks of the type specified (via the --memthreshold flag, which defaults to only reporting leaks related to explicit user code), nothing is printed.  Otherwise, leaked memory allocations are summarized according to the description, with each line in the report specifying how many items fit the description on that line and the total amount of memory leaked for all the items fitting that description. The lines are sorted in order from most to least total memory.
Most of this is implemented in Chapel.  An associative array is created to
aggregate the amount of memory for each unique description.  This array is then
copied to an arithmetic array, which is then sorted.  Finally, the array is
written out to stdout.

A couple .good files were changed since the output when using --memfinalstat is
now different.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15545 3a8e244f-b0f2-452b-bcba-4c88e055c3ca