This fix resolves some of the race conditions within the PVM comm layer.
This fix resolves some of the race conditions within the PVM comm layer. It
accomplishes this by changing when locks are taken out (removes the pack_lock
dictating when PVM packages can be put together or torn apart, but using
macros, it extends the scope of the pvm_lock to include all package work).
This change also completely breaks the exit with a hack to chpl_comm_barrier
which prevents the code from entering the chpl_exit_all barrier (thus
holding the lock), but it allows all the other code to run before hanging at
the end. A permanent fix may involve broadcasting a ready to exit status.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15547 3a8e244f-b0f2-452b-bcba-4c88e055c3ca