Made bool methods non-generic
Previously, a function defined def bool.foo() was treated as
defining the function for all bool widths.  This wasn't anything
particularly deep, I just implemented it the same as integers,
not fully realizing/remembering that def int.foo() creates a
function for all int widths.  David's test showed this to be
a problem given our inter-bool coercions, so I took out that
capability and his test works as expected.

The fact that I didn't remember that def int.foo() creates a
function for all widths is worrisome to me, and it seems
unintuitive given that int in any other context means "default
width int".  As a result, I'm going to look into where we rely
on this to see whether it can be removed.  Perhaps def
integral.foo() or somesuch could be used if the generic version
is required.  Steve points out that our parser doesn't accept
def int(32).foo(), so another exploration would be whether we
could accept this, either for the built-in scalar types or
more generally...



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15552 3a8e244f-b0f2-452b-bcba-4c88e055c3ca