Changed interpretation of "def scalartype.foo()"
For quite some time now, we have treated function definitions like:

  def int.foo()

as a definition of method foo() for all integer types.  In coming back
across this recently, it seemed very confusing to me given that "int"
typically means "default-width integer".  

The best rationale for this behavior that we could recall is that it
provides a way to define functions for non-default width primitive
types (e.g., int(16)) since the parse does not accept "def
int(16).foo()".  Yet this rationale seems weak since this capability
still does not permit a user to just define that single function (a
reasonable thing to want to do).

For this reason, I checked into where we currently rely on this
capability and found that the only place in the module code is in
defining the .re and .im methods for complex types.

This lead me to the following proposal:

- let's remove the current interpretation of what it means to define a
  function on "int" and just make it define the method for the default
  width int as it appears to be.

- to the extent that we want to support such a generic method
  capability in the future, let's do so by defining generic types
  (similar to "integral") and support the capability to define
  functions on those generic types as a way of creating the functions
  for all types that they encompass

- in a separate vein, we can explore ways to permit functions to be
  defined on non-standard datatypes like int(16) -- by special casing
  them in the parser or generalizing the parsing of function
  prototypes.

To this end, I disabled the capability as proposed in bullet one and
added a generic complex type "chpl_anycomplex" (not currently intended
for end-user use) in order to define the .re and .im methods.

This required updating one test -- test_integer_method3 -- that used
to test the old behavior and now demonstrates the correct behavior
under the new rules.

Steve and I couldn't find anywhere that the old behavior was
documented in the spec, so I haven't updated anything there.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15553 3a8e244f-b0f2-452b-bcba-4c88e055c3ca