This change alters how PVM launching begins and completes.
This change spawns the PVM daemon and uses C routines to begin the Chapel
work as opposed to hacking the shell to run programs. While the work isn't
complete (needs to be moved into chpl_launch), the framework is in place.
Right now, the user can manually start the xxx_real application from the
PVM daemon (still) or from the shell (new, as long as console is running),
but the launcher itself run from xxx doesn't work. Will need new printf
mechanism to correctly debug. Also note there's a mechanism for the comm
layer to signal back to the launcher that it is done (variable: end).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15568 3a8e244f-b0f2-452b-bcba-4c88e055c3ca