Improved parser dependences slightly to work toward supporting parallel make
This change improves the dependences on chapel.tab.h slightly so
that parallel make works better.  However, there are still higher
level problems in the form of assumptions about order in which
prerequisites get made that need to be dealt with before parallel
make will work with any certainty.  I have a prototype of this
in my directory but it's not ready for prime-time yet.  This
checkin represents a step toward better Makefiles in any case.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15570 3a8e244f-b0f2-452b-bcba-4c88e055c3ca