Removed the makechpl script and slimmed down the README.macosx file
We decided to remove the makechpl script due to (1) the fact that
it has not worked on the release tarball as documented for several
users and myself, (2) Sam is no longer here to develop and champion
it, (3) its speed benefits primarily come from recompilation which
most users won't need to do.

This made the README.macosx file fairly empty.  As discussed, I put
in mention of the Xcode project, a disclaimer that it will not be
kept up-to-date, and an advertisement for the community to help us
with this if interested.  Since README.macosx is not particularly
necessary for most users, I removed its pointer from our top-level
README file (it'd be more a distraction than a reasonable next step
for most users) but left it in the documentation directory's README
for completeness.

Removed the makechpl script from inclusion in the release via the
gen_release script.

Nothing like a milestone review to get some housecleaning done.  :)



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15583 3a8e244f-b0f2-452b-bcba-4c88e055c3ca