Adding fflush calls to stdout writelns
When running GASNet runs on multiple distinct locales (as with
CX1 and XT4), I've been noticing for quite awhile that we don't
get all the initial output to the console before we start running
the actual trials.  For example, the problem size for STREAM or RA
will be printed out but then nothing following that.  

Investigating, it appears that there were some fflush calls missing
from our standalone writeln() calls and that adding these fflush
calls causes all the output to be printed.  I'm not entirely sure
why the lack of these fflush calls was resulting in the past
behavior -- in particular why some things were printing out and not
others.  My guess is that it has to do with the buffer sizes used
by C in its use of [f]printf, but I'm not positive.  In any case,
this change appears to be an improvement (and I wish I had it in 
v0.8 for my demo today :).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15584 3a8e244f-b0f2-452b-bcba-4c88e055c3ca