This change cleans up some code and takes out two race conditions.
This change takes out two race conditions. First, on exit, it appeared after
completing a task, a locale would exit occasionally with an error, thus
hanging the app. It appears this was a result of incorrectly setting the
chpl_localeID. This value now lines up with the order it is added to the
group. Second, the launcher code had a bug on sending arguments via spawn().
There's a hack, now, to make sure argc remains the same for the main
routine, but it works as a way to make progress.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15588 3a8e244f-b0f2-452b-bcba-4c88e055c3ca