Removed memTableEntry::prevInstalled and memTableEntry::nextInstalled. There is no reason to keep the memory table entries in a linked list since they are already in a hash table.  This was only used for printing the table.  With this checkin, the table is now printed in sorted order.
I believe I can add a compressed version of this print now that it is
sorted to subsume the --memfinalstat flag.  That uses a Chapel hash
table over strings rather than the qsort so I'll want to make sure
that the new version is not slower.  I think it may be faster since
Chapel associative domains over strings have not yet been optimized.

Updated a few .good for printMemTable tests because I now print the
description before the address (so sorting is deterministic).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15592 3a8e244f-b0f2-452b-bcba-4c88e055c3ca