Add a future about let expressions and one about mixing types and values.
The let expression future causes the compiler to issue a warning about control
reaching the end of a function that returns a value. That is not the case.
Even if it were true, it wouldn't be very helpful since it is about a compiler
generated function.

The second future is about applying equality operators to type and values as
in: "var b = int == 1;".  I don't think should be defined.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15594 3a8e244f-b0f2-452b-bcba-4c88e055c3ca