Removed the deep copy of the description string in the memory entry used by memory tracking.  Since the description strings are constants in the code, we don't need to copy them.  This also makes the change to printMemTable easier since we can avoid string comparisons and rely instead on pointer comparisons.  My next step is to replace these strings with integers.  This improves performance measurable (40.8s on fft with n=14 down to 37.4s).  The timed portion is impacted due to faster memory tracking too (4.1s to 3.1s).
Added extra argument to printMemTable, bool aggregated, that is used
to signal that every entry with the same description should be grouped
together.  This grouping is done in O(d*n) time where d is the number
of unique descriptions and n is the number of memory entries.  Since d
is small, this has a similar complexity of the hash table
implementation in chpl_printMemTable that is implemented via an
associative domain of strings in the chapel_code runtime file.  It is
significantly faster (37.4s on fft with n=14 down to 14.6s).  The
timed portion is unaffected; this portion of this commit only impacts
the time to print the table.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15597 3a8e244f-b0f2-452b-bcba-4c88e055c3ca