Renamed all visible chplmemtrack.h functions so that they have a chpl_ prefix.  For example, printMemTable becomes chpl_printMemTable.
Added a primitive for chpl_printMemTable to allow users to call it via
this primitive.  It was previously externed in the standard Memory
module and the line number and filename were passed explicitly.  By
going through a primitive the line number and filename can be set
correctly.

Also added primitive for chpl_printMemStat and associated function
printMemStat in the Memory module.  Removed cases where this was
externed directly in tests (again with line number and filename being
passed in with bogus values).

Fixed line number in memory/shannon/memtrackFlag.chpl.  It is more
accurate now that we are passing these in correctly.

Updated line numbers in users/brodtkorb/coercionQuestion due to
changing ChapelBase.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15599 3a8e244f-b0f2-452b-bcba-4c88e055c3ca