Made memfinalstat, memstat, memtrace, and memtrack bools static to chplmemtrack.c.
Created new function chpl_reportMemInfo() to report memory information
at the end of a program run (stats via memstat and/or leaked memory
info via memfinalstat).  This is now called from chpl_exit_all and
unifies the two places this was done in the code before this change.

Replaced uses of __LINE__ and __FILE__ in the runtime (comm_armci,
comm_mpi, and comm_pvm) with 0 and 0 where these are passing on line
number and filename info that is reported to the user.  These line
numbers and filenames should always point to Chapel code.  Where they
do not, we typically pass in 0,0 to indicate that we don't have line
number or filename information.  The point in the runtime source is
not of interest to the user.  Instead, we should eliminate more 0,0
calls by passing line number and filename information to the runtime
in more places.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15600 3a8e244f-b0f2-452b-bcba-4c88e055c3ca