Changed memory tracking "description" argument from a string to an integer.  This impacts every call to chpl_malloc and chpl_realloc in the runtime.  There is now a new table chpl_rt_memDescs and a corresponding enumeration chpl_rt_enum_memDescs which can be used to add new memory description strings for new chpl_malloc or chpl_realloc calls in the runtime.  The compiler creates its own table of description strings.  The runtime function chpl_memDescString takes an integer and returns the corresponding description string.
This approach has a number of advantages.  We no longer need to copy
the description string in compiler code, ever.  We don't have to copy
the description string when running on other locales.  We can do a
faster comparison.  We eliminate the need for a hash table (and
string/int hashing) when printing out the leaked memory table.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15606 3a8e244f-b0f2-452b-bcba-4c88e055c3ca