This fix corrects how args are passed to pvm_spawn via the launcher.
This fix corrects how args are passed to pvm_spawn via the launcher. Prior
to this change, the args were copied directly, and it assumed only -nl <num>
was passed. Now it allows for any number of arguments. It also appends
the number of locations to the end so the comm layer grabs it for its use.
The comm layer appropriately strips it after it uses it.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15608 3a8e244f-b0f2-452b-bcba-4c88e055c3ca