Refactored and reorganized memory tracking code some more.
Optimized routine to add table entries.  It used to do an unnecessary
lookup.  This has a significant impact on performance (14.6s on fft
with n=14 down to 1.0s--no overhead over running the benchmark without
--memreport).  Handling of chpl_free does require a lookup so there
are still performance issues; the hash table really needs to be
resized when it fills up.

Renamed memfinalstat variables to memreport to reflect the changed
name of the flag.

Enabled memoryUsed() and printMemStat() with just --memtrack flag,
since statistics are now gathered with this flag.  The flag --memstat
is now only used to print out the statistics.  Removed the hack that
squelched this printing if either of the above mentioned functions
were called.  Updated execopts files for many tests.

This checkin also includes many other minor reorganizational changes.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15615 3a8e244f-b0f2-452b-bcba-4c88e055c3ca