Refactored the launcher code so that it needs to include less of the runtime source files.  In particular, it no longer includes chplexit.c, chpl_mem.c, chplmemtrack.c and gdb.c.  In addition, it no longer includes code in mem-*/.
Divided header files chplcomm, chplexit, chpl_mem, chplmemtrack and chplrt
(similarly to chplthreads) so that they define more functions for the runtime
than for the launcher.  For the launcher, we use macros instead.  For example;
for the launcher, chpl_malloc is defined as a function, but for the runtime,
chpl_malloc is prototyped and implemented in chpl_mem.c.

Removed some #ifndef LAUNCHER statements from some runtime source code in
chpl_mem.c and chplmemtrack.c.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15620 3a8e244f-b0f2-452b-bcba-4c88e055c3ca