This change fixes the exit from the PVM comm to be more graceful.
This change fixes the exit from the PVM comm to be more graceful. It turns
out, pvm_halt basically nukes the world, and it causes an internal segfault
signal to be passed around s.t. the test scripts don't work correctly (even
though they output what they are supposed to). After this change, processes
use pvm_exit (and in the case of the launcher, it just lets the code
continue). We have to make sure the daemons get properly cleaned up, however,
so we kill them in the shell.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15622 3a8e244f-b0f2-452b-bcba-4c88e055c3ca