Added back support, with semantic changes, for a memory tracking threshold and for a memory log file.  The high-level of these changes are that the threshold now applies to the memory tracking itself, which should improve performance when only tracking large allocations, and thus every aspect of memory tracking and reporting; and that the log file now applies to all memory tracking and reporting output.
Added support for memory tracking of multi-locale programs.  Memory
tracking now happens on all locales, not just locale 0.  Memory
statistics are now printed in table form where there is a row for
every locale.  Verbose memory output is now reported on all locales.
If a memory log file is specified and the program is run on multiple
locales, there is a separate log file for each locale and these are
distinguished by appending a dot followed by the locale id to each
file.

Replaced memory tracking flags with configuration constants.  Added
the following configuration constants:

  config const
    memTrack: bool = false,
    memStats: bool = false, 
    memLeaks: bool = false,
    memMax: uint(64) = 0,
    memThreshold: uint(64) = 0,
    memLog: string = "";

Updated the documentation to reflect the change from flags to
configuration variables, the change in names, and several semantic
changes.  The new semantics are detailed as follows:

  --memTrack        : turns on memory tracking and enables the
                      functions memoryUsed, printMemStat, and
                      printMemTable.  This flag is implied by
                      memStats, memLeaks, and memMax.

  --memStats        : makes the call printMemStat() when the
                      program completes, printing out the following
                      statistics: current memory allocated (would be 0
                      if there were no leaks), maximum memory
                      allocated at the same time, total allocated
                      memory, and total freed memory.

  --memLeaks        : prints out a table of all memory that has not
                      been freed when the program completes.  It is
                      similar to printMemTable() except that the
                      entries in the table are aggregated and
                      summarized.

  --memMax=uint(64) : sets a maximum level of allocatable memory.  If
                      this constraint is violated, the program is
                      terminated.  If unset or 0, there is no
                      simulated maximum.

  --memThreshold=uint(64) : sets a minimum threshold for memory
                            tracking.  Allocations and frees that are
                            smaller than this threshold are ignored.
                            The default is 0.  Note that if memory
                            tracking is off, verbose memory output may
                            include frees that are smaller than this
                            threshold.  Alternatively, if memory
                            tracking is on, the output may omit frees
                            that are larger than this threshold.  This
                            is because the size of the free is only
                            known by looking at the table of tracked
                            memory.

  --memLog=string :   specifies a file where memory reporting is
                      redirected.  It is used to redirect the output
                      generated by printMemTable, by verbose memory
                      reporting (enabled by startVerboseMem and/or
                      startVerboseMemHere), by --memStats, and by
                      --memLeaks.  If memLog is unspecified or the
                      empty string, output is directed to standard
                      out.  For multi-locale runs, a dot and the
                      locale number is appended to the name of the
                      memLog so that each locale creates its own log
                      file.

Added primitive and runtime function chpl_setMemFlags that is used to
communicate the values in the configuration constants to variables
declared in the runtime.  It is called on all locales in the standard
module function chpl_startTrackingMemory which is inserted at the
beginning of the implicit or explicit main function in the Chapel
program.

Added support for setting configuration constants and variables of
bool type by omitting an initialization expression.  Thus --memTrack
is equivalent to --memTrack=true.  I believe this is a different
approach from revision 9441 that created this short-hand for
configuration parameters via the -s flag.  In that case, the value is
toggled.  At least with the configurations constants and variables,
which default to false, it seems like specifying the flag ought to set
it to true and specifying it twice ought to still set it to true.  Do
we want to support a --no-"config" to set bool constants and variables
to false?

This short-hand support also has some issues with the changes in
revision 15526 in which the = was made optional.  If --memTrack=true
is equivalent to --memTrack, then --memTrack true is interpreted as
setting memTrack to true and then passing argument true, resulting in
an error.  How should this be resolved?

Removed a bit of code to deal with the memory tracking flags in the
runtime.  Satisfying!

This commit also removes some error checking from our memory tracking
implementation, namely, checking for frees or reallocations of memory
that was never allocated.  There are a number of reasons why this is
untenable though the pressing reason with this commit is the change in
tracking only memory above a threshold; this means that a free or
realloc could just not be tracked.  Moving forward, as we free more
memory, it also seems that we may start to free memory that is
allocated before tracking was turned on.  Moreover, if we ever want to
turn tracking on or off at user-defined points in a program, then it
is very likely that we may free memory within this code section that
was allocated after it.

Updated many execopts files to deal with the name changes from flags
to configuration constants.

Updated several tests to make the output more independent of where
memory tracking is turned on (though it is still turned on in the same
spot).  Moving forward, it may make more sense to turn on tracking as
soon as possible rather than as close to user code as possible.

Updated some error messages that changed with the flags to
configuration constants change.

Removed two tests of memory checking during memory tracking.

Futurized memory/shannon/memstatEquals test which uses an equal sign
but does not specify an initialization.  This speaks to the open issue
mentioned above about allowing bool config constants to omit an
initialization and all config constants to omit the = sign.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15624 3a8e244f-b0f2-452b-bcba-4c88e055c3ca